email_triage_rl_hackathon/pyproject.toml
[project]
name = "email-triage-rl-hackathon"
version = "0.1.0"
description = "Email Triage RL Environment — Meta x PyTorch OpenEnv Hackathon"
requires-python = ">=3.10"

dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "openai",
    "requests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
server = "email_triage_rl_hackathon.server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["email_triage_rl_hackathon"]
